InspectorAction test dependency fix
moves inspector action tests to their own file so they can be setup on their own so as to not depend on the CreateMenu tests
